requirements.txt
aiomysql==0.2.0
alembic==1.13.1
black==24.4.2
cryptography==42.0.8
fastapi==0.111.0
greenlet==3.0.3
httptools==0.6.1
pydantic-settings==2.3.3
pytest==8.2.2
PyYAML==6.0.1
SQLAlchemy-Utils==0.41.2
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
